Added ClassLibrary to have a DLL Because this is mandatory for the assignment.
