+++ Project NMBS/MainWindow.xaml.cs
@@ -37,6 +37,7 @@ using transit_realtime;
 
 using Fishezzz;
 using System.Threading;
+using ExtensionsNMBS;
 
 namespace Project_NMBS
 {
