+++ Project NMBS/RouteExtensions.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static Fishezzz.Exceptions;
 
 namespace Fishezzz
 {
